refactor(reachability): keep the import-name aliases in package_utils
The flat src/agent_bom namespace is frozen, so the alias table moves into
package_utils, which already owns package-name normalization, instead of
adding a new top-level module.
